Teste projeto/requirements.txt
opencv-python
mediapipe
numpy
requests
python-dotenv
